Simple game using a while loop and if function
start/stop/quit functions to do basic stuff, added a fuel limit of 5. Cannot start after starting again, cannot drive if not started, cannot stop if already stopped etc. 
Clearly a lot of code was repeated, am looking to see if there is a way to reduce the number of lines.